refactor(storage): Decompose database/core.rs into focused modules
Extract concerns from core.rs (976 lines) into focused submodules:

- config.rs: DatabaseConfig, SpillPolicy, DEFAULT_COLUMNAR_CACHE_BUDGET
- constructors.rs: new(), with_path(), with_config(), Clone, Default
- cache.rs: Columnar cache integration methods
- index_ops.rs: Index and spatial index operations
- debug.rs: Debug utilities (debug_info, dump_tables)
- session.rs: SQL mode, session variables, security/role management

Results:
- core.rs reduced from 976 to 319 lines (well under 400 line target)
- All public APIs maintained via re-exports in mod.rs
- All 233 tests pass
- Clone semantics preserved (new cache for cloned database)

Closes #2705
