+++ crates/vibesql-storage/src/database/core.rs
@@ -4,192 +4,34 @@
 
 use super::lifecycle::Lifecycle;
 use super::metadata::Metadata;
-use super::operations::{Operations, SpatialIndexMetadata};
+use super::operations::Operations;
 use super::transactions::TransactionChange;
-use super::DatabaseConfig;
 use crate::columnar_cache::ColumnarCache;
 use crate::{QueryBufferPool, Row, StorageError, Table};
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::sync::Arc;
-use vibesql_ast::IndexColumn;
 
 pub use super::operations::SpatialIndexMetadata as ExportedSpatialIndexMetadata;
-
-/// Default columnar cache budget (256MB)
-const DEFAULT_COLUMNAR_CACHE_BUDGET: usize = 256 * 1024 * 1024;
 
 /// In-memory database - manages catalog and tables through focused modules
 #[derive(Debug)]
 pub struct Database {
     /// Public catalog access for backward compatibility
     pub catalog: vibesql_catalog::Catalog,
-    lifecycle: Lifecycle,
-    metadata: Metadata,
-    operations: Operations,
+    pub(super) lifecycle: Lifecycle,
+    pub(super) metadata: Metadata,
+    pub(super) operations: Operations,
     pub tables: HashMap<String, Table>,
     /// SQL compatibility mode (MySQL, SQLite, etc.)
-    sql_mode: vibesql_types::SqlMode,
+    pub(super) sql_mode: vibesql_types::SqlMode,
     /// Buffer pool for reducing query execution allocations
-    query_buffer_pool: QueryBufferPool,
+    pub(super) query_buffer_pool: QueryBufferPool,
     /// LRU cache for columnar table representations
     /// Shared via Arc to allow cloning without duplicating cache data
-    columnar_cache: Arc<ColumnarCache>,
-}
-
-impl Clone for Database {
-    fn clone(&self) -> Self {
-        Database {
-            catalog: self.catalog.clone(),
-            lifecycle: self.lifecycle.clone(),
-            metadata: self.metadata.clone(),
-            operations: self.operations.clone(),
-            tables: self.tables.clone(),
-            sql_mode: self.sql_mode.clone(),
-            query_buffer_pool: self.query_buffer_pool.clone(),
-            // Clone creates a new cache with same config but empty data
-            // This is intentional - cloned databases shouldn't share cache state
-            columnar_cache: Arc::new(ColumnarCache::new(self.columnar_cache.max_memory())),
-        }
-    }
+    pub(super) columnar_cache: Arc<ColumnarCache>,
 }
 
 impl Database {
-    /// Create a new empty database
-    ///
-    /// Note: Security is disabled by default for backward compatibility with existing code.
-    /// Call `enable_security()` to turn on access control enforcement.
-    pub fn new() -> Self {
-        Database {
-            catalog: vibesql_catalog::Catalog::new(),
-            lifecycle: Lifecycle::new(),
-            metadata: Metadata::new(),
-            operations: Operations::new(),
-            tables: HashMap::new(),
-            sql_mode: vibesql_types::SqlMode::default(),
-            query_buffer_pool: QueryBufferPool::new(),
-            columnar_cache: Arc::new(ColumnarCache::new(DEFAULT_COLUMNAR_CACHE_BUDGET)),
-        }
-    }
-
-    /// Create a new database with a specific storage path
-    ///
-    /// The provided path will be used as the root directory for database files.
-    /// Index files will be stored in `<path>/data/indexes/`.
-    ///
-    /// # Example
-    /// ```rust
-    /// use std::path::PathBuf;
-    /// use vibesql_storage::Database;
-    ///
-    /// let db = Database::with_path(PathBuf::from("/var/lib/myapp/db"));
-    /// // Index files will be stored in /var/lib/myapp/db/data/indexes/
-    /// ```
-    pub fn with_path(path: PathBuf) -> Self {
-        let mut db = Self::new();
-        db.operations.set_database_path(path.join("data"));
-        db
-    }
-
-    /// Create a new database with a specific configuration
-    ///
-    /// Allows setting memory budgets, disk budgets, and spill policy for adaptive
-    /// index management.
-    ///
-    /// # Example
-    /// ```rust
-    /// use vibesql_storage::{Database, DatabaseConfig};
-    ///
-    /// // Browser environment with limited memory
-    /// let db = Database::with_config(DatabaseConfig::browser_default());
-    ///
-    /// // Server environment with abundant memory
-    /// let db = Database::with_config(DatabaseConfig::server_default());
-    /// ```
-    pub fn with_config(config: DatabaseConfig) -> Self {
-        let columnar_cache_budget = config.columnar_cache_budget;
-        let mut db = Self::new();
-        db.sql_mode = config.sql_mode.clone();
-        db.columnar_cache = Arc::new(ColumnarCache::new(columnar_cache_budget));
-        db.operations.set_config(config);
-        db
-    }
-
-    /// Create a new database with both path and configuration
-    ///
-    /// # Example
-    /// ```rust
-    /// use std::path::PathBuf;
-    /// use vibesql_storage::{Database, DatabaseConfig};
-    ///
-    /// let db = Database::with_path_and_config(
-    ///     PathBuf::from("/var/lib/myapp/db"),
-    ///     DatabaseConfig::server_default()
-    /// );
-    /// ```
-    pub fn with_path_and_config(path: PathBuf, config: DatabaseConfig) -> Self {
-        let columnar_cache_budget = config.columnar_cache_budget;
-        let mut db = Self::new();
-        db.sql_mode = config.sql_mode.clone();
-        db.columnar_cache = Arc::new(ColumnarCache::new(columnar_cache_budget));
-        db.operations.set_database_path(path.join("data"));
-        db.operations.set_config(config);
-        db
-    }
-
-    /// Create a new database with both path and configuration (async version for WASM)
-    ///
-    /// This async version is required for WASM to properly initialize OPFS storage
-    /// without blocking the event loop.
-    ///
-    /// # Example
-    /// ```rust,no_run
-    /// # use std::path::PathBuf;
-    /// # use vibesql_storage::{Database, DatabaseConfig};
-    /// # async fn example() {
-    /// let db = Database::with_path_and_config_async(
-    ///     PathBuf::from("/vibesql-data"),
-    ///     DatabaseConfig::browser_default()
-    /// ).await.unwrap();
-    /// # }
-    /// ```
-    #[cfg(target_arch = "wasm32")]
-    pub async fn with_path_and_config_async(
-        path: PathBuf,
-        config: DatabaseConfig,
-    ) -> Result<Self, crate::StorageError> {
-        let columnar_cache_budget = config.columnar_cache_budget;
-        let mut db = Self::new();
-        db.sql_mode = config.sql_mode.clone();
-        db.columnar_cache = Arc::new(ColumnarCache::new(columnar_cache_budget));
-        db.operations.set_database_path(path.join("data"));
-        db.operations.set_config(config);
-
-        // Initialize OPFS storage asynchronously
-        db.operations.init_opfs_async().await?;
-
-        Ok(db)
-    }
-
-    /// Reset the database to empty state (more efficient than creating a new instance).
-    ///
-    /// Clears all tables, resets catalog to default state, and clears all indexes and transactions.
-    /// Useful for test scenarios where you need to reuse a Database instance.
-    /// Preserves database configuration (path, storage backend, memory budgets) across resets.
-    pub fn reset(&mut self) {
-        self.catalog = vibesql_catalog::Catalog::new();
-        self.lifecycle.reset();
-        self.metadata = Metadata::new();
-
-        // Reset operations in place to preserve database_path, storage backend, and config
-        self.operations.reset();
-
-        self.tables.clear();
-
-        // Clear the columnar cache
-        self.columnar_cache.clear();
-    }
-
     // ============================================================================
     // Transaction Management
     // ============================================================================
@@ -437,321 +279,12 @@ impl Database {
     }
 
     // ============================================================================
-    // Debug and Utilities
-    // ============================================================================
-
-    /// Get debug information about database state
-    pub fn debug_info(&self) -> String {
-        let mut output = String::new();
-        output.push_str("=== Database Debug Info ===\n");
-        output.push_str(&format!("Tables: {}\n", self.list_tables().len()));
-        for table_name in self.list_tables() {
-            if let Some(table) = self.get_table(&table_name) {
-                output.push_str(&format!(
-                    "  - {} ({} rows, {} columns)\n",
-                    table_name,
-                    table.row_count(),
-                    table.schema.column_count()
-                ));
-            }
-        }
-        output
-    }
-
-    /// Dump all table contents in readable format
-    pub fn dump_tables(&self) -> String {
-        let mut output = String::new();
-        for table_name in self.list_tables() {
-            if let Ok(dump) = self.dump_table(&table_name) {
-                output.push_str(&dump);
-                output.push('\n');
-            }
-        }
-        output
-    }
-
-    /// Dump a specific table's contents
-    pub fn dump_table(&self, name: &str) -> Result<String, StorageError> {
-        let table =
-            self.get_table(name).ok_or_else(|| StorageError::TableNotFound(name.to_string()))?;
-
-        let mut output = String::new();
-        output.push_str(&format!("=== Table: {} ===\n", name));
-
-        let col_names: Vec<String> = table.schema.columns.iter().map(|c| c.name.clone()).collect();
-        output.push_str(&format!("{}\n", col_names.join(" | ")));
-        output.push_str(&format!("{}\n", "-".repeat(col_names.join(" | ").len())));
-
-        for row in table.scan() {
-            let values: Vec<String> = row.values.iter().map(|v| format!("{}", v)).collect();
-            output.push_str(&format!("{}\n", values.join(" | ")));
-        }
-
-        output.push_str(&format!("({} rows)\n", table.row_count()));
-        Ok(output)
-    }
-
-    // ============================================================================
-    // Security and Role Management
-    // ============================================================================
-
-    /// Set the current session role for privilege checks
-    pub fn set_role(&mut self, role: Option<String>) {
-        self.lifecycle.set_role(role);
-    }
-
-    /// Get the current session role (defaults to "PUBLIC" if not set)
-    pub fn get_current_role(&self) -> String {
-        self.lifecycle
-            .current_role()
-            .map(|s| s.to_string())
-            .unwrap_or_else(|| "PUBLIC".to_string())
-    }
-
-    /// Check if security enforcement is enabled
-    pub fn is_security_enabled(&self) -> bool {
-        self.lifecycle.is_security_enabled()
-    }
-
-    /// Disable security checks (for testing)
-    pub fn disable_security(&mut self) {
-        self.lifecycle.disable_security();
-    }
-
-    /// Enable security checks
-    pub fn enable_security(&mut self) {
-        self.lifecycle.enable_security();
-    }
-
-    // ============================================================================
-    // Session Variables
-    // ============================================================================
-
-    /// Set a session variable (MySQL-style @variable)
-    pub fn set_session_variable(&mut self, name: &str, value: vibesql_types::SqlValue) {
-        self.metadata.set_session_variable(name, value);
-    }
-
-    /// Get a session variable value
-    pub fn get_session_variable(&self, name: &str) -> Option<&vibesql_types::SqlValue> {
-        self.metadata.get_session_variable(name)
-    }
-
-    /// Clear all session variables
-    pub fn clear_session_variables(&mut self) {
-        self.metadata.clear_session_variables();
-    }
-
-    // ============================================================================
-    // SQL Mode
-    // ============================================================================
-
-    /// Get the current SQL compatibility mode
-    pub fn sql_mode(&self) -> vibesql_types::SqlMode {
-        self.sql_mode.clone()
-    }
-
-    /// Set the SQL compatibility mode at runtime
-    ///
-    /// This allows changing the SQL dialect (MySQL, SQLite, etc.) during a session.
-    /// The `@@sql_mode` session variable is automatically updated to reflect the change.
-    ///
-    /// # Example
-    /// ```rust
-    /// use vibesql_storage::Database;
-    /// use vibesql_types::SqlMode;
-    ///
-    /// let mut db = Database::new();
-    /// assert!(matches!(db.sql_mode(), SqlMode::MySQL { .. }));
-    ///
-    /// db.set_sql_mode(SqlMode::SQLite);
-    /// assert!(matches!(db.sql_mode(), SqlMode::SQLite));
-    /// ```
-    pub fn set_sql_mode(&mut self, mode: vibesql_types::SqlMode) {
-        self.sql_mode = mode.clone();
-
-        // Update the @@sql_mode session variable to reflect the new mode
-        let mode_string = match &mode {
-            vibesql_types::SqlMode::MySQL { flags } => {
-                // Build MySQL mode string from flags
-                let mut modes = Vec::new();
-                if flags.strict_mode {
-                    modes.push("STRICT_TRANS_TABLES");
-                }
-                if flags.pipes_as_concat {
-                    modes.push("PIPES_AS_CONCAT");
-                }
-                if flags.ansi_quotes {
-                    modes.push("ANSI_QUOTES");
-                }
-                // Add common MySQL defaults if no specific flags are set
-                if modes.is_empty() {
-                    "NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_ENGINE_SUBSTITUTION".to_string()
-                } else {
-                    modes.join(",")
-                }
-            }
-            vibesql_types::SqlMode::SQLite => "SQLITE".to_string(),
-        };
-
-        self.metadata.set_session_variable(
-            "SQL_MODE",
-            vibesql_types::SqlValue::Varchar(mode_string),
-        );
-    }
-
-    // ============================================================================
     // Query Buffer Pool
     // ============================================================================
 
     /// Get a reference to the query buffer pool for reusing allocations
     pub fn query_buffer_pool(&self) -> &QueryBufferPool {
         &self.query_buffer_pool
-    }
-
-    // ============================================================================
-    // Index Management
-    // ============================================================================
-
-    /// Create an index
-    pub fn create_index(
-        &mut self,
-        index_name: String,
-        table_name: String,
-        unique: bool,
-        columns: Vec<IndexColumn>,
-    ) -> Result<(), StorageError> {
-        self.operations.create_index(
-            &self.catalog,
-            &self.tables,
-            index_name,
-            table_name,
-            unique,
-            columns,
-        )
-    }
-
-    /// Check if an index exists
-    pub fn index_exists(&self, index_name: &str) -> bool {
-        self.operations.index_exists(index_name)
-    }
-
-    /// Get index metadata
-    pub fn get_index(&self, index_name: &str) -> Option<&super::indexes::IndexMetadata> {
-        self.operations.get_index(index_name)
-    }
-
-    /// Get index data
-    pub fn get_index_data(&self, index_name: &str) -> Option<&super::indexes::IndexData> {
-        self.operations.get_index_data(index_name)
-    }
-
-    /// Update user-defined indexes for update operation
-    pub fn update_indexes_for_update(
-        &mut self,
-        table_name: &str,
-        old_row: &Row,
-        new_row: &Row,
-        row_index: usize,
-    ) {
-        self.operations.update_indexes_for_update(
-            &self.catalog,
-            table_name,
-            old_row,
-            new_row,
-            row_index,
-        );
-    }
-
-    /// Update user-defined indexes for delete operation
-    pub fn update_indexes_for_delete(&mut self, table_name: &str, row: &Row, row_index: usize) {
-        self.operations
-            .update_indexes_for_delete(&self.catalog, table_name, row, row_index);
-    }
-
-    /// Rebuild user-defined indexes after bulk operations that change row indices
-    pub fn rebuild_indexes(&mut self, table_name: &str) {
-        self.operations
-            .rebuild_indexes(&self.catalog, &self.tables, table_name);
-    }
-
-    /// Drop an index
-    pub fn drop_index(&mut self, index_name: &str) -> Result<(), StorageError> {
-        self.operations.drop_index(index_name)
-    }
-
-    /// List all indexes
-    pub fn list_indexes(&self) -> Vec<String> {
-        self.operations.list_indexes()
-    }
-
-    /// List all indexes for a specific table
-    pub fn list_indexes_for_table(&self, table_name: &str) -> Vec<String> {
-        self.operations.list_indexes_for_table(table_name)
-    }
-
-    // ============================================================================
-    // Spatial Index Methods
-    // ============================================================================
-
-    /// Create a spatial index
-    pub fn create_spatial_index(
-        &mut self,
-        metadata: SpatialIndexMetadata,
-        spatial_index: crate::index::SpatialIndex,
-    ) -> Result<(), StorageError> {
-        self.operations.create_spatial_index(metadata, spatial_index)
-    }
-
-    /// Check if a spatial index exists
-    pub fn spatial_index_exists(&self, index_name: &str) -> bool {
-        self.operations.spatial_index_exists(index_name)
-    }
-
-    /// Get spatial index metadata
-    pub fn get_spatial_index_metadata(&self, index_name: &str) -> Option<&SpatialIndexMetadata> {
-        self.operations.get_spatial_index_metadata(index_name)
-    }
-
-    /// Get spatial index (immutable)
-    pub fn get_spatial_index(&self, index_name: &str) -> Option<&crate::index::SpatialIndex> {
-        self.operations.get_spatial_index(index_name)
-    }
-
-    /// Get spatial index (mutable)
-    pub fn get_spatial_index_mut(&mut self, index_name: &str) -> Option<&mut crate::index::SpatialIndex> {
-        self.operations.get_spatial_index_mut(index_name)
-    }
-
-    /// Get all spatial indexes for a specific table
-    pub fn get_spatial_indexes_for_table(
-        &self,
-        table_name: &str,
-    ) -> Vec<(&SpatialIndexMetadata, &crate::index::SpatialIndex)> {
-        self.operations.get_spatial_indexes_for_table(table_name)
-    }
-
-    /// Get all spatial indexes for a specific table (mutable)
-    pub fn get_spatial_indexes_for_table_mut(
-        &mut self,
-        table_name: &str,
-    ) -> Vec<(&SpatialIndexMetadata, &mut crate::index::SpatialIndex)> {
-        self.operations.get_spatial_indexes_for_table_mut(table_name)
-    }
-
-    /// Drop a spatial index
-    pub fn drop_spatial_index(&mut self, index_name: &str) -> Result<(), StorageError> {
-        self.operations.drop_spatial_index(index_name)
-    }
-
-    /// Drop all spatial indexes associated with a table (CASCADE behavior)
-    pub fn drop_spatial_indexes_for_table(&mut self, table_name: &str) -> Vec<String> {
-        self.operations.drop_spatial_indexes_for_table(table_name)
-    }
-
-    /// List all spatial indexes
-    pub fn list_spatial_indexes(&self) -> Vec<String> {
-        self.operations.list_spatial_indexes()
     }
 
     // ============================================================================
@@ -782,195 +315,5 @@ impl Database {
     /// Clear all cached procedure/function bodies
     pub fn clear_routine_cache(&mut self) {
         self.metadata.clear_routine_cache();
-    }
-
-    // ============================================================================
-    // Columnar Cache Methods
-    // ============================================================================
-
-    /// Get columnar representation of a table, using cache if available
-    ///
-    /// This method provides an Arc-wrapped columnar representation of the table,
-    /// enabling zero-copy sharing between queries. The cache automatically manages
-    /// memory via LRU eviction.
-    ///
-    /// # Arguments
-    /// * `table_name` - Name of the table to get columnar representation for
-    ///
-    /// # Returns
-    /// * `Ok(Some(Arc<ColumnarTable>))` - Cached or newly converted columnar data
-    /// * `Ok(None)` - Table not found
-    /// * `Err(StorageError)` - Conversion failed
-    ///
-    /// # Example
-    /// ```rust,ignore
-    /// if let Some(columnar) = db.get_columnar("lineitem")? {
-    ///     // Use columnar data for SIMD operations
-    /// }
-    /// ```
-    pub fn get_columnar(&self, table_name: &str) -> Result<Option<Arc<crate::ColumnarTable>>, StorageError> {
-        // Check cache first
-        if let Some(cached) = self.columnar_cache.get(table_name) {
-            return Ok(Some(cached));
-        }
-
-        // Table not in cache - need to get table and convert
-        let table = match self.get_table(table_name) {
-            Some(t) => t,
-            None => return Ok(None),
-        };
-
-        // Convert to columnar format
-        let columnar = table.scan_columnar()?;
-
-        // Insert into cache and return
-        let cached = self.columnar_cache.insert(table_name, columnar);
-        Ok(Some(cached))
-    }
-
-    /// Invalidate columnar cache entry for a table
-    ///
-    /// Called automatically when a table is modified (INSERT/UPDATE/DELETE)
-    /// to ensure the cache doesn't serve stale data.
-    pub fn invalidate_columnar_cache(&self, table_name: &str) {
-        self.columnar_cache.invalidate(table_name);
-    }
-
-    /// Clear all columnar cache entries
-    pub fn clear_columnar_cache(&self) {
-        self.columnar_cache.clear();
-    }
-
-    /// Get columnar cache statistics
-    ///
-    /// Returns statistics about cache hits, misses, evictions, and conversions.
-    /// Useful for monitoring cache effectiveness and tuning the cache budget.
-    pub fn columnar_cache_stats(&self) -> crate::columnar_cache::CacheStats {
-        self.columnar_cache.stats()
-    }
-
-    /// Get current columnar cache memory usage in bytes
-    pub fn columnar_cache_memory_usage(&self) -> usize {
-        self.columnar_cache.memory_usage()
-    }
-
-    /// Get columnar cache memory budget in bytes
-    pub fn columnar_cache_budget(&self) -> usize {
-        self.columnar_cache.max_memory()
-    }
-
-    /// Set the columnar cache memory budget
-    ///
-    /// Note: This creates a new cache, discarding all cached data.
-    /// Call this before loading data for best results.
-    pub fn set_columnar_cache_budget(&mut self, max_bytes: usize) {
-        self.columnar_cache = Arc::new(ColumnarCache::new(max_bytes));
-    }
-}
-
-impl Default for Database {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use vibesql_types::{MySqlModeFlags, SqlMode, SqlValue};
-
-    #[test]
-    fn test_set_sql_mode_changes_mode() {
-        let mut db = Database::new();
-
-        // Default is MySQL
-        assert!(matches!(db.sql_mode(), SqlMode::MySQL { .. }));
-
-        // Change to SQLite
-        db.set_sql_mode(SqlMode::SQLite);
-        assert!(matches!(db.sql_mode(), SqlMode::SQLite));
-
-        // Change back to MySQL
-        db.set_sql_mode(SqlMode::MySQL {
-            flags: MySqlModeFlags::default(),
-        });
-        assert!(matches!(db.sql_mode(), SqlMode::MySQL { .. }));
-    }
-
-    #[test]
-    fn test_set_sql_mode_updates_session_variable() {
-        let mut db = Database::new();
-
-        // Set to SQLite mode
-        db.set_sql_mode(SqlMode::SQLite);
-
-        // Check session variable reflects the change
-        let sql_mode_var = db.get_session_variable("SQL_MODE");
-        assert!(sql_mode_var.is_some());
-        if let Some(SqlValue::Varchar(mode_str)) = sql_mode_var {
-            assert_eq!(mode_str, "SQLITE");
-        } else {
-            panic!("Expected SQL_MODE to be a Varchar");
-        }
-    }
-
-    #[test]
-    fn test_set_sql_mode_mysql_with_flags() {
-        let mut db = Database::new();
-
-        // Set MySQL with specific flags
-        db.set_sql_mode(SqlMode::MySQL {
-            flags: MySqlModeFlags {
-                pipes_as_concat: true,
-                ansi_quotes: true,
-                strict_mode: true,
-            },
-        });
-
-        // Check session variable contains the flags
-        let sql_mode_var = db.get_session_variable("SQL_MODE");
-        assert!(sql_mode_var.is_some());
-        if let Some(SqlValue::Varchar(mode_str)) = sql_mode_var {
-            assert!(mode_str.contains("STRICT_TRANS_TABLES"));
-            assert!(mode_str.contains("PIPES_AS_CONCAT"));
-            assert!(mode_str.contains("ANSI_QUOTES"));
-        } else {
-            panic!("Expected SQL_MODE to be a Varchar");
-        }
-    }
-
-    #[test]
-    fn test_set_sql_mode_mysql_default_flags() {
-        let mut db = Database::new();
-
-        // Set MySQL with default flags (all false)
-        db.set_sql_mode(SqlMode::MySQL {
-            flags: MySqlModeFlags::default(),
-        });
-
-        // Check session variable has default MySQL modes
-        let sql_mode_var = db.get_session_variable("SQL_MODE");
-        assert!(sql_mode_var.is_some());
-        if let Some(SqlValue::Varchar(mode_str)) = sql_mode_var {
-            // Default should include common MySQL defaults
-            assert!(mode_str.contains("NO_ZERO_IN_DATE") || mode_str.contains("NO_ENGINE_SUBSTITUTION"));
-        } else {
-            panic!("Expected SQL_MODE to be a Varchar");
-        }
-    }
-
-    #[test]
-    fn test_sql_mode_affects_subsequent_queries() {
-        let mut db = Database::new();
-
-        // Start in MySQL mode
-        assert!(matches!(db.sql_mode(), SqlMode::MySQL { .. }));
-
-        // Switch to SQLite
-        db.set_sql_mode(SqlMode::SQLite);
-
-        // Verify the mode is available for query execution
-        let mode = db.sql_mode();
-        assert!(matches!(mode, SqlMode::SQLite));
     }
 }
